requirements.txt
Cython==3.1.2
huggingface-hub==0.16.4
Jinja2==3.1.6
joblib==1.5.1
matplotlib==3.10.3
nltk==3.9.1
numpy==2.2.6
openai-clip==1.0.1
pandas==2.3.0
pillow==11.2.1
ruclip==0.0.2
safetensors==0.5.3
scikit-learn==1.7.0
sentence-transformers==2.3.1
tokenizers==0.14.1
torch==2.7.1
torchmetrics==1.7.2
torchvision==0.22.1
tqdm==4.67.1
transformers==4.35.0
youtokentome==1.0.6